requirements.txt
Flask==2.3.3
torch==2.0.1
transformers==4.35.2
tokenizers==0.15.0
numpy==1.24.3
scipy==1.11.4
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
lxml==4.9.3
flask-cors==4.0.0
